move components to folder and split state into new file
